+++ system/cron.php
@@ -19,11 +19,10 @@ ORM::configure('logging', true);
 include "autoload/Hookers.php";
 
 // notification message
-if (file_exists("uploads/notifications.json")) {
-    $_notifmsg = json_decode(file_get_contents('uploads/notifications.json'), true);
-} else {
-    $_notifmsg = json_decode(file_get_contents('uploads/notifications.default.json'), true);
+if(file_exists("system/uploads/notifications.json")){
+    $_notifmsg =json_decode(file_get_contents('system/uploads/notifications.json'), true);
 }
+$_notifmsg_default = json_decode(file_get_contents('system/uploads/notifications.default.json'), true);
 
 //register all plugin
 foreach (glob("plugin/*.php") as $filename) {
@@ -59,12 +58,15 @@ $result = ORM::for_table('tbl_appconfig')->find_many();
 foreach ($result as $value) {
     $config[$value['setting']] = $value['value'];
 }
+
+$_c = $config;
+
 date_default_timezone_set($config['timezone']);
 
-$textExpired = $_notifmsg['expired'];
+$textExpired = Lang::getNotifText('expired');
 
-$d = ORM::for_table('tbl_user_recharges')->where('status', 'on')->find_many();
-
+$d = ORM::for_table('tbl_user_recharges')->where('status', 'on')->where('expiration', date("Y-m-d"))->find_many();
+echo "Found ".count($d)." user(s)\n";
 run_hook('cronjob'); #HOOK
 
 foreach ($d as $ds) {
@@ -95,12 +97,20 @@ foreach ($d as $ds) {
                     if (Package::rechargeUser($ds['customer_id'], $p['routers'], $p['id'], 'Customer', 'Balance')) {
                         // if success, then get the balance
                         Balance::min($ds['customer_id'], $p['price']);
+                        echo "plan enabled: $p[enabled] | User balance: $c[balance] | price $p[price]\n";
+                        echo "auto renewall Success\n";
                     } else {
+                        echo "plan enabled: $p[enabled] | User balance: $c[balance] | price $p[price]\n";
+                        echo "auto renewall Failed\n";
                         Message::sendTelegram("FAILED RENEWAL #cron\n\n#u$c[username] #buy #Hotspot \n" . $p['name_plan'] .
                             "\nRouter: " . $router_name .
                             "\nPrice: " . $p['price']);
                     }
+                }else{
+                    echo "no renewall | plan enabled: $p[enabled] | User balance: $c[balance] | price $p[price]\n";
                 }
+            }else{
+                echo "no renewall | balance $config[enable_balance] auto_renewal $c[auto_renewal]\n";
             }
         } else echo " : ACTIVE \r\n";
     } else {
@@ -131,7 +141,7 @@ foreach ($d as $ds) {
                         // if success, then get the balance
                         Balance::min($ds['customer_id'], $p['price']);
                     } else {
-                        Message::sendTelegram("FAILED RENEWAL #cron\n\n#u$c[username] #buy #Hotspot \n" . $p['name_plan'] .
+                        Message::sendTelegram("FAILED RENEWAL #cron\n\n#u$c[username] #buy #PPPOE \n" . $p['name_plan'] .
                             "\nRouter: " . $router_name .
                             "\nPrice: " . $p['price']);
                     }