+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 # CHANGELOG
 
+## 2023.8.24
+
+- Balance transfer between Customer
+- Optimize Cronjob
+- View Customer Info
+- Ajax for select customer
+
+## 2023.8.18
+
+- Fix Auto Renewall Cronjob
+- Add comment to Mikrotik User
+
 ## 2023.8.16
 
 - Admin Can Add Balance to Customer